shop_project/requirements.txt
asgiref==3.8.1
django==4.2.17
sqlparse==0.5.1
tzdata==2024.2
djangorestframework==3.15.2
drf-yasg==1.21.8
psycopg2-binary==2.9.10
django-environ==0.11.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
pyjwt==2.10.1
django-ckeditor==6.7.2
python-decouple==3.8
django-recaptcha==4.0.0
requests==2.32.3
drf-spectacular>=0.28.0
python-telegram-bot==21.10
python-dotenv==1.0.1
django-cors-headers==4.6.0
